refactor: create player positions module
