+++ src/components/forms/player/players-filter-form.tsx
@@ -15,7 +15,7 @@ import { TeamBasicDataDto } from '@/modules/teams/types'
 import { CompetitionGroupBasicDataDto } from '@/modules/competition-groups/types'
 import { CompetitionBasicDataDto } from '@/modules/competitions/types'
 import { CountryDto } from '@/modules/countries/types'
-import { PlayerPositionDto } from '@/types/player-positions'
+import { PlayerPositionDto } from '@/modules/player-positions/types'
 
 import { Container } from '../container'
 import { FilterFormActions } from '../filter-form-actions'
