+++ src/pages/players/create.tsx
@@ -4,9 +4,9 @@ import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 import { CreatePlayerForm } from '@/components/forms/player/create-player'
 import { Loader } from '@/components/loader/loader'
 import { PageHeading } from '@/components/page-heading/page-heading'
-import { usePlayerPositionsList } from '@/lib/player-positions'
 import { withSessionSsr } from '@/lib/session'
 import { useCountriesList } from '@/modules/countries/hooks'
+import { usePlayerPositionsList } from '@/modules/player-positions/hooks'
 import { useCreatePlayer } from '@/modules/players/hooks'
 import { useTeamsList } from '@/modules/teams/hooks'
 import { redirectToLogin } from '@/utils/redirect-to-login'
